Fixed failing test and made change-locale filenames consistent.
