+++ src/components/choose-locale/change-locale.component.scss
