+++ src/components/choose-locale/change-locale.component.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import Select from 'carbon-components-react/es/components/Select';
 import SelectItem from 'carbon-components-react/es/components/SelectItem';
-import styles from './changelocal.component.scss';
+import styles from './change-locale.component.scss';
 import { PostUserProperties } from '../../offline/user-properties.resource';
 import { refetchCurrentUser } from '@openmrs/esm-framework';
 import { LoggedInUser } from '../../types';
